add actor factory for creating actors
