+++ calvin/runtime/north/actormanager.py
@@ -14,13 +14,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from calvin.actorstore.store import ActorStore
 from calvin.utilities import dynops
 from calvin.runtime.south.plugins.async import async
 from calvin.utilities.calvinlogger import get_logger
 from calvin.utilities.calvin_callback import CalvinCB
 import calvin.utilities.calvinresponse as response
-from calvin.actor.actor import ShadowActor
+from calvin.actor.actor_factory import ActorFactory
 
 _log = get_logger(__name__)
 
@@ -34,10 +33,11 @@ class ActorManager(object):
 
     """docstring for ActorManager"""
 
-    def __init__(self, node):
+    def __init__(self, node, factory=None):
         super(ActorManager, self).__init__()
         self.actors = {}
         self.node = node
+        self.factory = factory if factory else ActorFactory(node)
 
     def new(self, actor_type, args, state=None, prev_connections=None, connection_list=None, callback=None,
             signature=None):
@@ -55,15 +55,8 @@ class ActorManager(object):
         """
         _log.debug("class: %s args: %s state: %s, signature: %s" % (actor_type, args, state, signature))
         a = self._new(actor_type, args, state, signature)
-
-        if prev_connections:
-            # Convert prev_connections to connection_list format
-            connection_list = self._prev_connections_to_connection_list(prev_connections)
-
         self.node.control.log_actor_new(a.id, a.name, actor_type)
-        if connection_list:
-            # Migrated actor
-            self.connect(a.id, connection_list, callback=callback)
+        self._setup_connections(a, prev_connections=prev_connections, connection_list=connection_list)
 
         if callback:
             callback(status=response.CalvinResponse(True), actor_id=a.id)
@@ -78,21 +71,11 @@ class ActorManager(object):
         """
         _log.analyze(self.node.id, "+", {'actor_type': actor_type, 'state': state})
 
-        try:
-            if state:
-                a = self._create_from_state(actor_type, state)
-            else:
-                a = self._create_new(actor_type, args)
-        except Exception as e:
-            _log.exception("Actor creation failed")
-            raise(e)
-
-        # Store the actor signature to enable GlobalStore lookup
-        a.signature_set(signature)
+        a = self.factory.create_actor(actor_type=actor_type, state=state, args=args, signature=signature)
 
         self.actors[a.id] = a
-
         self.node.storage.add_actor(a, self.node.id)
+
         return a
 
     def new_replica(self, actor_type, args, prev_connections, callback):
@@ -128,75 +111,6 @@ class ActorManager(object):
                 translated_connection_list.append((node_id, port_id_translations[port_id], peer_node_id, peer_port_id))
 
         return translated_connection_list
-
-    def _new_actor(self, actor_type, actor_id=None):
-        """Return a 'bare' actor of actor_type, raises an exception on failure."""
-        (found, is_primitive, class_) = ActorStore().lookup(actor_type)
-        if not found:
-            # Here assume a primtive actor, now become shadow actor
-            _log.analyze(self.node.id, "+ NOT FOUND CREATE SHADOW ACTOR", {'class': class_})
-            found = True
-            is_primitive = True
-            class_ = ShadowActor
-        if not found or not is_primitive:
-            _log.error("Requested actor %s is not available" % (actor_type))
-            raise Exception("ERROR_NOT_FOUND")
-        try:
-            # Create a 'bare' instance of the actor
-            a = class_(actor_type, actor_id=actor_id)
-        except Exception as e:
-            _log.exception("")
-            _log.error("The actor %s(%s) can't be instantiated." % (actor_type, class_.__init__))
-            raise(e)
-        try:
-            a._calvinsys = self.node.calvinsys()
-            a.check_requirements()
-        except Exception as e:
-            _log.exception("Catched new from state")
-            _log.analyze(self.node.id, "+ FAILED REQS CREATE SHADOW ACTOR", {'class': class_})
-            a = ShadowActor(actor_type, actor_id=actor_id)
-            a._calvinsys = self.node.calvinsys()
-        return a
-
-    def _create_new(self, actor_type, args):
-        """Return an initialized actor in PENDING state, raises an exception on failure."""
-        try:
-            a = self._new_actor(actor_type)
-            # Now that required APIs are attached we can call init() which may use the APIs
-            human_readable_name = args.pop('name', '')
-            a.name = human_readable_name
-            self.node.pm.add_ports_of_actor(a)
-            a.init(**args)
-            a.setup_complete()
-        except Exception as e:
-            _log.exception(e)
-            raise(e)
-        return a
-
-    def _create_from_state(self, actor_type, state):
-        """Return a restored actor in PENDING state, raises an exception on failure."""
-        try:
-            _log.analyze(self.node.id, "+", state)
-            a = self._new_actor(actor_type, actor_id=state['id'])
-            if '_shadow_args' in state:
-                # We were a shadow, do a full init
-                args = state.pop('_shadow_args')
-                state['_managed'].remove('_shadow_args')
-                a.init(**args)
-                # If still shadow don't call did_migrate
-                did_migrate = isinstance(a, ShadowActor)
-            else:
-                did_migrate = True
-            # Always do a set_state for the port's state
-            a._set_state(state)
-            self.node.pm.add_ports_of_actor(a)
-            if did_migrate:
-                a.did_migrate()
-            a.setup_complete()
-        except Exception as e:
-            _log.exception("Catched new from state %s %s" % (a, dir(a)))
-            raise(e)
-        return a
 
     def destroy(self, actor_id):
         # @TOOD - check order here
@@ -333,10 +247,8 @@ class ActorManager(object):
             state = actor.state()
             self.destroy(actor.id)
             self.node.proto.actor_new(node_id, callback, actor_type, state, ports)
-        else:
-            # FIXME handle errors!!!
-            if callback:
-                callback(status=status)
+        elif callback:  # FIXME handle errors!!!
+            callback(status=status)
 
     def replicate(self, actor_id, node_id, callback=None):
         """Replicate an actor actor_id to peer node node_id """
@@ -373,6 +285,15 @@ class ActorManager(object):
             for in_id in in_list:
                 cl.append((self.node.id, out_port_id, in_id[0], in_id[1]))
         return cl
+
+    def _setup_connections(self, actor, prev_connections=None, connection_list=None, callback=None):
+        if prev_connections:
+            # Convert prev_connections to connection_list format
+            connection_list = self._prev_connections_to_connection_list(prev_connections)
+
+        if connection_list:
+            # Migrated actor
+            self.connect(actor.id, connection_list, callback=callback)
 
     def connect(self, actor_id, connection_list, callback=None):
         """
